requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
certifi==2020.6.20
Django==3.0.3
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
wincertstore==0.2
